인스타그램 모듈 분리
