+++ dependencies.gradle
@@ -1,0 +1,27 @@
+ext {
+    versions = [
+            'minSdkVersion'    : 21,
+            'compileSdkVersion': 29,
+            'targetSdkVersion' : 29,
+            'gradle'           : '3.5.0',
+            'kotlin'           : '1.3.50'
+    ]
+
+    deps = [
+            'rx'      : [
+                    'rxJava'   : 'io.reactivex.rxjava2:rxjava:2.2.8',
+                    'rxAndroid': 'io.reactivex.rxjava2:rxandroid:2.1.1',
+            ],
+
+            'androidx': [
+                    'appcompat': 'androidx.appcompat:appcompat:1.1.0',
+                    'core': 'androidx.core:core-ktx:1.1.0'
+            ],
+
+            'glide'   : [
+                    'glide'     : 'com.github.bumptech.glide:glide:4.9.0',
+                    'annotation': 'com.github.bumptech.glide:compiler:4.9.0',
+            ],
+
+    ]
+}
