+++ myinatagram/build.gradle
@@ -1,4 +1,4 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
@@ -8,12 +8,8 @@ android {
     //buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "com.tistory.blackjin.storeapplicaion"
         minSdkVersion versions.minSdkVersion
         targetSdkVersion versions.targetSdkVersion
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     dataBinding {
@@ -34,26 +30,18 @@ android {
 }
 
 dependencies {
-    implementation project(':myinatagram')
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${versions.kotlin}"
 
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation deps.androidx.appcompat
     implementation deps.androidx.core
-    
+
     implementation deps.glide.glide
+    implementation 'androidx.appcompat:appcompat:1.1.0'
     annotationProcessor deps.glide.annotation
 
     implementation deps.rx.rxJava
     implementation deps.rx.rxAndroid
-
-    implementation 'com.jakewharton.timber:timber:4.7.1'
-    implementation 'gun0912.ted:tedimagepicker:1.0.9'
-
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test:runner:1.2.0'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
 }
