+++ myinatagram/src/main/java/com/tistory/blackjin/myinatagram/InstagramActivity.kt
@@ -1,0 +1,12 @@
+package com.tistory.blackjin.myinatagram
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class InstagramActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_instagram)
+    }
+}
